+++ client.go
@@ -1,7 +1,7 @@
 package glaw
 
 import (
-    "context"
+	"context"
 	"encoding/json"
 	"fmt"
 	"net/http"
@@ -13,24 +13,24 @@ import (
 )
 
 type Client struct {
-    client http.Client
-    token string
+	client http.Client
+	token  string
 }
 
 type Option interface {
-    apply(*Client) (*Client, error)
+	apply(*Client) (*Client, error)
 }
 
 func NewClient(opts ...Option) (c *Client, err error) {
-    c := &Client{}
+	c := &Client{}
 
-    for _, opt := range opts {
-        if c, err := opt.apply(c); err != nil {
-            return c, err
-        }
-    }
+	for _, opt := range opts {
+		if c, err := opt.apply(c); err != nil {
+			return c, err
+		}
+	}
 
-    return c, err
+	return c, err
 }
 
 func (c *Client) Do(req *http.Request) (resp *http.Response, err error) {
@@ -49,44 +49,44 @@ func (c *Client) Do(req *http.Request) (resp *http.Response, err error) {
 }
 
 func ProcessResponse(resp *http.Response, to interface{}) error {
-    defer resp.Body.Close()
-    return json.NewDecoder(resp.Body).Decode(to)
+	defer resp.Body.Close()
+	return json.NewDecoder(resp.Body).Decode(to)
 }
 
 // /lol/platform/v3/champion-rotations
 type Request struct {
-    Method string
-    Domain string
-    Version Version
-    Region Region
-    Uri string
-    Body io.Reader
+	Method  string
+	Domain  string
+	Version Version
+	Region  Region
+	Uri     string
+	Body    io.Reader
 }
 
 func (r Request) NewHttpRequestWithCtx(ctx context.Context) *http.Request {
-    template := `https://%s/lol/%s/%s/%s`
-    u := fmt.Sprintf(template, r.Region.Base(), Domain, Version, Uri)
+	template := `https://%s/lol/%s/%s/%s`
+	u := fmt.Sprintf(template, r.Region.Base(), Domain, Version, Uri)
 
-    return http.NewRequestWithContext(ctx, r.Method, u, r.Body)
+	return http.NewRequestWithContext(ctx, r.Method, u, r.Body)
 }
 
 func (r Request) NewHttpRequest() *http.Request {
-    ctx := context.Background()
+	ctx := context.Background()
 
-    return r.NewHttpRequestWithCtx(ctx)
+	return r.NewHttpRequestWithCtx(ctx)
 }
 
 func NewRequest(method string, region Region, version Version) {
-   return Request {
-       Method: method,
-       Version: version,
-       Region: region,
-    }
+	return Request{
+		Method:  method,
+		Version: version,
+		Region:  region,
+	}
 }
 
 type Version string
 
 const (
-    V3 Version = `v3`
-    V4 Version = `v4`
+	V3 Version = `v3`
+	V4 Version = `v4`
 )
