--- v3/champions/champions.go
@@ -1,7 +1,0 @@
-package champions
-
-type ChampionInfo struct {
-	FreeChampionIds              []int `json:"freeChampionIds"`
-	FreeChampionIdsForNewPlayers []int `json:"freeChampionIdsForNewPlayers"`
-	MaxNewPlayerLevel            int   `json:"maxNewPlayerLevel"`
-}
