--- v4/mastery/mastery.go
@@ -1,13 +1,0 @@
-package mastery
-
-type ChampionMastery struct {
-	ChampionLevel                int    `json:"championLevel"`
-	ChestGranted                 bool   `json:"chestGranted"`
-	ChampionPoints               int    `json:"championPoints"`
-	ChampionPointsSinceLastLevel int    `json:"championPointsSinceLastLevel"`
-	ChampionPointsUntilNextLevel int    `json:"championPointsUntilNextLevel"`
-	SummonerID                   string `json:"summonerId"`
-	TokensEarned                 int    `json:"tokensEarned"`
-	ChampionID                   int    `json:"championId"`
-	LastPlayTime                 int64  `json:"lastPlayTime"`
-}
